+++ lib/mixins/bouncing_object.dart
@@ -2,8 +2,16 @@ import 'package:bonfire/bonfire.dart';
 
 // Mixin responsable to give the bounce behavior. (experimental)
 mixin BouncingObject on BlockMovementCollision {
-  double bouncingReflectFactor = 1.0;
-  Vector2? currentCenter;
+  double _bouncingReflectFactor = 1.0;
+  bool _bouncingObjectEnabled = true;
+
+  void setupBouncingObject({
+    bool enabled = true,
+    double reflectFactor = 1.0,
+  }) {
+    _bouncingObjectEnabled = enabled;
+    _bouncingReflectFactor = reflectFactor;
+  }
 
   bool onBouncingCollision(PositionComponent other) {
     return other is! Sensor;
@@ -15,14 +23,18 @@ mixin BouncingObject on BlockMovementCollision {
     Direction? direction,
     Vector2 lastDisplacement,
   ) {
-    if (onBouncingCollision(other) && !velocity.isZero()) {
+    if (onBouncingCollision(other) &&
+        !velocity.isZero() &&
+        _bouncingObjectEnabled) {
       if (direction == Direction.left || direction == Direction.right) {
-        velocity.x = velocity.x * -bouncingReflectFactor;
+        velocity.x = velocity.x * -_bouncingReflectFactor;
       } else if (direction == Direction.up || direction == Direction.down) {
-        velocity.y = velocity.y * -bouncingReflectFactor;
+        velocity.y = velocity.y * -_bouncingReflectFactor;
       } else {
         stopMove();
       }
+    } else {
+      super.onBlockedMovement(other, direction, lastDisplacement);
     }
   }
 }